requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
openpyxl>=3.1.0
